Replace only exactly two newlines with paragraph separators
If there's a string of at least two newlines in the middle of
non-newline stuff, replace the first two of them with a paragraph
separator. Other newlines are replaced by `<br/>`.

Always wrap in `<p>` even when there's only one line (so we won't have
to count lines via expensive string splitting).